feat(Main): add personalInfo state to Main component
The state is shared between the EditPersonalInfoPanel and
PreviewPersonalInfoPanel components, hence it is defined in the Main
(the closest common component), and lifted up through EditPanel and
PreviewPanel components.

personalInfo is initialized as an object, with a dummy test property,
just test that the state is properly passed to such components.