style(PreviewPage): add style to PreviewPage component
Only the "top-level" style is defined. The style of the specific
components used in there component are styled in their own module.

The aspect ratio has been set to that of an A4 paper. In this div,
however, the content will not be split into multiple pages. A scroll
will be enabled.

The CSS property container-type has been set to inline-size. This way,
it is possible to use sizes relative to this page size in the
descendants elements. Although, note that older browser will not be
supported.