feat(PreviewItemWithDates, PreviewListWithDates): setup PreviewItemWithDates and PreviewListWithDates components
PreviewItemWithDates initialize an item with some data displayed as Text
and a range of dates attached. The actual data and the properties to
show are decided based on the inupt parameter.

PreviewListWithDates creates a list of  PreviewItemWithDates items,
based on the provided parameter.

These will be used in the PreviewEducationPanel and
PreviewProfessionalExperience panel.