refactor(body, Main): edit the styles of this components for responsiveness
The body is set to occupy the full viewport size: any overflow, without
any scrolling.

Instead, the Main component now always position its flex items in a row:
the horizontal scroll is set to auto, so that if the EditPanel and
PreviewPanel do not fit horizontally, a horizontal scrollbar appears.