feat(useIsOverflow): add useIsOverflow() React custom hook
This custom hooks is used to check if an element has overflow.

The element is specified by ref parameter, which in turn is a reference
to an element created with useRef() hook.

The parameter isVerticalOverflow specifies whether the vertical or
horizontal overflow is to be checked.

The preCallback and postCallback callbacks can be specified, to perform
an action before checking the overflow condition, or after that,
respectively.

If the browser supports ResizeObserver, the overflow condition is
checked again.

see: https://stackoverflow.com/a/70909891
see: https://www.robinwieruch.de/react-custom-hook-check-if-overflow/