feat(EditListSelectionButtons): add EditListSelectionButtons component
This component initializes a list of buttons.

It assumes there is an external list of items, each of which has a .id
property. This list is passed as a prop.

Another prop specifies the id of the currently selected item, while a
prop to set it is also passed. Then, for each item, there is a button
which selects the corresponding item. Only one button can be
selected at a time.

A 'new item' button is also added, which sets the selected item as an
empty item with id=-1 (assuming ids start from 0).