feat(EditEducationPanel): add input elements to edit the "degree", "institution", "address" fields
When selecting a specific item button (#0, #1,#2, ...), a form is shown
which is initalized with the data associated with the item.

When selecting the new item button, an empty form is shown. When the
user starts typing anything, a new item is actually added to the
education array. This prevents adding too many empty items.

An helper module has been setup to handle the operations on a generic
array of items, where each item is an object with an id field, just like
the education one.