feat(DateInputWithLabel): add DateInputWithLabel component
This component contains two selectInput component, one to select a
month, and one to select a year.

A range can be specified for both the possible months and years through
the firstYear, lastYear, firstMonth, and lastMonth props.

Other props that can be specified include: id, name, valueMonth and
valueYear (the current controlled month a year), setMonth and setYear
(to set such values), a required flag, and a label.

Note that a single label is used for both the month and year, as they
are essetntially both part of a label. Both the month and year select
elements are however labeled using "aria-label" property.

This component is different from the <input type="month"/> element,
because the month can be specified as empty, selecting in fact just a
year.