feat(DateInputWithLabel): add DateInputWithLabel component
This component contains two selectInput component, one to select a
month, and one to select a year.

An optional range can be specified, providing the values for the minimum
date (fromMonth, fromYear) and maximum date (toMonth, toYear).

The intervals for both the possible months and years are then computed
based on the minimum, maximum, and selected dates, computing the
firstYear, lastYear, firstMonth, and lastMonth variables. If some of
these  values are missing, some default values are used (1900,
currentYear, January, and December, respectively).

Note that those values are computed so that it is not possible to select
an invalid date. Eg, if:
    - the minimum date is March 2022
    - the maximum date is October 2024
    - the current date is February 2023
we get:
    - firstYear: 2023
    - lastYear: 2024
    - firstMonth: January
    - lastMonth: December
Recall that you cannot change the month and the year simultaneously, so:
    - any month is valid if you change the month (keeping year 2023)
    - year 2022 would not be selectable (keeping month February),
      because February 2022 is out of range.

Other props that can be specified include: id, name, valueMonth and
valueYear (the current controlled month a year), setMonth and setYear
(to set such values), a required flag, and a label.

Note that a single label is used for both the month and year, as they
are essetntially both part of a label. Both the month and year select
elements are however labeled using "aria-label" property.

This component is different from the <input type="month"/> element,
because the month can be specified as empty, selecting in fact just a
year.