feat(IntervalDateInputWithLabels): add IntervalDateInputWithLabels component
This component contains two DateInputWithLabel components, one for a
"from date" and one for a "to date", that let you specify an interval of
dates.

Date are specified with a year and an optional month.

The date ranges (fromMonth, fromYear, toMonth, toYear) are automatically
computed based on the current selected date.

Other props that can be specified include: fromDateProp and toDateProp
string variables which specify the suffix of the names of the props for
the two dates (note that a 'Month' or 'Year' string will be appended),
and the inputProps object, item object and setValueFor function (just
like the ArrayOfInputWithLabel component).