fix(itemsArrayManagement, EditPersonalInfoPanel): use updater functions when updating a state which is an object
This prevents the object state not being properly updated when multiple
state updates are setted in the same rendering, because the most recent
copy of the state is used instead of the one of the previous render.

This has been done both in the EditPersonalInfoPanel component and in
the itemsArrayManagement functions.

For the itemsArrayManagement functions, some methods have been renamed
and the array input has been removed from the interface of some
functions (the updated copy of the array is obtained by the first
argument of the updater function).

The itemsArrayManagement functions used in EditEducationPanel component
have been properly edited to reflect these changes.