feat!: flatten React states from Main component (1/3 - edit inputProperties.json)
Instead of having nested arrays in education and professionalExperience
React states from Main component, a less nested structure has been used
based on:

https://medium.com/free-code-camp/optimising-the-state-shape-of-your-react-app-with-redux-3a280e6ef436

(although Redux is not used here).

This allows for a simplified handling of the updates for the nested
properties.

In this first phase, only the inputProperties.json file has been
flattened.

BREAKING CHANGE: the preview and edit panels are now broken. They will
be adapted to the next state structure in the next commits.