feat(OtherInfoListInputWithLabel, EditEducationPanel): add OtherInfoListInputWithLabel component
The component is used to edit the optional other information for the
selected Education item. The handling of each info resemble the handling
of the EditEducationPanel component; in particular, there are:

- a EditListSelectionButtons component to select the info to edit;
- a EditItemPositionInList component to move or delete the currently
  selected info;
- two InputWithLabel components to edit the heading and the text of the
  currently selected info.

The allIdsPath and byIdPath arrays to properly access the nested
education state have been properly constructed based on the input props.

The style will be propely applied next.

The component has been instantiated in the EditEducationPanel component,
having updated inputProperties.json with the new required properties.

Note that there is a bug that is to be fixed in a future commit, by
making the initialization of a new item as soon as the + button is
clicked. This is particularly important for the otherInfo group of
inputs. As implemented now, when a new item is added (which is empty,
without an id), editing the otherInfo portion first breaks the execution
(because the 'education' item is not actually present in the education
state yet, and the search/insertion of a nested new property fails).
For now, this otherInfo group of input is shown only if the item to be
edited is not empty (ie, it actually exists), which avoids the problem.

Minor changes have been applied to EditListSelectionButtons.jsx and
EditItemPositionInList.css.