feat(DownloadButton): add DownloadButton component
A button that downloads an HTML element as a PDF when clicked is
implemented in its own module, and instantiated in the Main component.

To fix:
- embed custom fonts in the PDF to use it
- convert svg icons to png (they are not saved properly)

Issues:
Sometimes, when exporting the PDF, the block elements aren't split
correctly between pages, despite having used the *page-break-inside:
avoid;* and *break-inside: avoid;* css properties.