fix(PreviewPersonalInfoPanel): use png icons instead of svg icons, to export them correctly with jsPdf
The Icon component has been properly modified to load a png image, if
required, which can be correctly exported with jsPdf.

The involved icons have been converted to png, too.

Note that the icon color cannot be modified if it is png, so the icon
png image must be in the correct color (eg, white).