fix(DownloadButton): fix the downloaded pdf not containing the whole content in multi-page view
The fix was using 'overflow:visible' for the .preview-page to be
exported to pdf.

Normally, this property is set to 'overflow: hidden auto'.

Then, as .preview-page gets the .printing class when the export is in
progress, that property opportunely set when that class is applied.

Before this commit, these changes applied with the .printing class were
temporarily visible. Now, this gets hidden: the 'onclone' property of
html2canvas has been set through the jsPdf interface, to apply the
.printing class only on the cloned element which that library uses
internally to generate the pdf, which is not visible.