feat(ToggleDarkThemeButton): add ToggleDarkThemeButton
The button toggles the dark theme of the app. A custom .dark-theme class
is set on a container div of the app, based on the value of the new
darkThemeSet state. The initial value of this state is based on the
system preferences.

The button has been instantiated in the Header component.