feat(Main): add personalization state defining the preview accent color
The personalization state has been defined in the Main component. As of
now this is an object which only contains the accent color to be used
in the preview page.

While not implemented yet, it will be possible to let the user chose
this accent color from the editor user interface (instead of just hard
coding it in the variable.css file).

Other properties will be added in the personalization object next.