feat(Main): add personalization state defining the preview page main font
Altough not implemented yet, this will allow the user to change the
font used in the preview page.

Both the CSS variable used to set the font in the .preview-page page and
the fonts to be included in the exported pdf (through the Download
Button) are now set based on this state variable.