feat(App, Main): implement persistence to state variables
The 7 state variables defined in Main component holding the data to
create the cv page are now saved in local storage, for data persistence
when the user refresh the page.

The state variables are now initialized with the values from local
storage (if present), or otherwise with some default data.

useEffect has been used to update the data stored in localStorage
whenever each state variable changes.

The same has been done for the darkThemeSet state variable in App
component. Indeed, if the user refresh the page, the same theme scheme
is re-applied.