feat(EditPersonalInfo): add input for optional photo in EditPersonalInfoPanel component
The optional photo is converted to base64 when provided, and the
corresponding state variable is updated, which contains a property with
an object with such base64 data, the original filename, date of last
edit and type.

When rendering the input, if a photo is already present in the
localStorage or the initial state, the base64 data is converted to a
blob which is used to initialize the corresponding input.

Note that the previewPersonalInfoPanel does not render such image yet.