feat(App): disable some app buttons when the displayed form is invalid.
The disabled buttons are ".download-btn", ".print-btn",
".edit-panel-button", ".edit-list-selection-new-item-button", and
".edit-list-selection-item-button".

Then, when the form is invalid, the user cannot export or print it, nor
it can change the currently displayed edit form.

To do that, a single event listener has been set to the App component
outer div, on the click event: it has been set to record the capture
phase, too. Then, when clicking on a button, this event listener is
called before the button executes its action.

In this event listener, first it is checked that the clicking point is
in an actual button. If this is the case, it is checked whether that
button is a button to be disabled when the displayed form is invalid.

Eg, assume the form is invalid: we would disable the download button to
not let the user download a cv with invalid data, but we would not
disable the button that toggles the dark theme, or the button that
reset the cv.

Finally, the validity of each displayed form is checked and as soon an
invalid form is found, the propagation of the captured event is stopped
to prevent the functionality callback of the event listener to be
executed.