feat(PreviewPersonalInfoPanel): transform contact texts to hyperlinks
The contacts items have been implemented using the
AnchorWithTextWithIcon component.

The corresponding href depends on the property and state value (***)
associated with it:

- if 'email', a 'mailto:***' url is used;

- if 'phone', a 'tel:***' url is used;

- if 'location', 'https://maps.apple.com/maps?q=***' url is used. Note
  that this is redirected to 'https://www.google.com/maps?q=***'
  on non-Apple devices;

- otherwise, "***" is used as web url. If the protocol is not specified,
  https:// is used.

Note that as of now, the pdf obtained with the download button (using
jspdf) will ignore such links, while the pdf obtained with the print
button (using the save as pdf option) will keep such links.