fix(EditPanel): reset currentPanel when the cv data are re-initialized
The currentPanel state of EditPanel and the currentItemId state of
EditEducationPanel, EditProfessionalExperiencePanel, EditSkillsPanel,
EditLanguagesPanel, EditCoursesAndCertificatesPanel, and
EditOtherInfoListInputWithLabel were not properly reset when the user
either cleared the form data or loaded sample data, resulting in the
wrong form being displayed.

Then, when this happens, currentPanel has now been initialized to the
initial personal info panel (which does not have lists within it, and so
no currentItemId state to reset in there).

To do this, now the currentPanel state definition has been lifted up to
the App component, where it is defined through the useDefineDataStates
custom hook.

In useDefineDataStates, currentPanel has been set whenever setting new
data (through setDataFrom).